Fix MLP light & allow bidirectional diffuse
In some cases we want to allow the reflectance to have both directions, so add a bidirectional
flag for that case. In addition, allow the MLP to be indexed with some embedding so it can have
different values per scene (which are related).

In the case it's not bidiretional, clamp to 0.