Add joint all learned & const shadow
Adds a new kind of occlusion model, which is a combination of the fully learned model as well as
a raycasting one. This should provide better guarantees, as it will defer to the fully learned
model as it learns components, and as it progresses it should be able to instead rely directly
on raycasting for learning correct shadows. In addition, add a method to convert into this model
from an existing "all-learned" occlusion. It's probably not necessary to have a direct analytic
component at the start since it will mostly be wrong anyway.