Fix up DNeRF & Add better NeRV test
Previously, DNeRF was not updated to integrate with replacing the time component, so added that
in. As well as predicting splines instead of just using an MLP directly.

In addition, the NeRV test set was annoying since it required separate runs. It now integrates
so it can be run in the same iteration.