Fix intermediate size & impl de casteljau's
Intermediate size only worked for VolSDF before, because I had accidentally conflated
total_latent_size which should be the input for the density function with the intermediate
vector that is passed from the density model to the reflectance model.

This now fixes it so that an actual latent vector is passed from normal models to the
reflectance models.