Add in a positional refl with gamma correction
Allowing a reflectance model to vary any color between views is too strong of an assumption.
Instead, let's assume we can have a gamma correction when viewing from different angles. I'm
not sure whether this is a reasonable assumption or not, but I feel it is fairly reasonable in
that it allows for varying light intensity.