Allow dyn --> refl latent parameters
In certain cases, we are interested in passing information at a given timestep about the
illumination of a point. I.e, if lighting changes, we may want to be able to pass that from a
dynamic model into the reflectance. This is dangerous in that it may permit models to learn too
much at a specific timestep, but may be useful in that it only allows _recoloring_ over time.

Have to see how well it works!