Fix up static cam focal + LongDynNeRF
Also remove loaders devices. Now, the static camera will allow for passing a guess at the wide
angle for a scene, altho it may not be correct. In addition, we now leave the frames on the CPU,
and send them to the GPU during training which should free up a bit of memory, esp for video
reconstruction. Remove some trained parameters on the camera which were not being optimized
properly, and also finish an attempt at LongDynamicNeRF with splines, fixing bugs along the way.

It will take a few days to train the entire model, but have to see if it works.